[FIX] industry_fsm: new quotation creation
- Create a new quotation from field service;
- Change the salesperson in the quotation (to send an invitation email).

Before this commit, an error was raised. This occurs because in the
function that creates the new quotation, the 'draft' value was specified
as default for the status field (in the context). This is correct, when
creating the sale.order model, but it generates an error when creating
the mail.mail model that don't accept 'draft' as value for its status
field.

Now, the 'draft' value for the status field was removed from the
context, this doesn't change the behaviour because in the model
sale.order the default for the status field is already the 'draft'
value.

opw-2177751

closes odoo/enterprise#7934

X-original-commit: bbdb2fbd050eac960472242f2048dc9ba0308af2
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>