[FIX] documents: record selection responsiveness
This commit increases the visual responsiveness of the selection
of documents by moving the rendering of the `DocumentsInspector` to
a different callStack so it can be painted after the selection.

The slowness of the rendering of the `DocumentsInspector` is caused
by the fact that we currently destroy and re-create the `DocumentsInspector`
as we change the record selection (instead of updating it).

closes odoo/enterprise#7911

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>