[IMP] stock_barcode, *: editable location
In the picking barcode view and the picking batch barcode view, the user
can click on the source/destination location to change it.
A list with sublocations will be displayed and then the user can click
on one of them to change the src/dest location of the page's move lines.

task-2123726