[FIX] mrp_plm: BoM and Routing changes
- Create a BOM for product A
- Create an ECO for the created BOM
- Start New Revision
- Add a product to the BOM

2 operations 'Add' are recorded on the ECO instead of one.

The core of the issue is the use of a computed One2many field:
`bom_change_ids`. The compute methods retrieves the differences between
the old and the new BOM, and adds it to the list. However, it's
impossible to make sure the compute method will be called only once in a
transaction. And this is what causes the issue: the compute method is
called twice, so the changes are recorded twice.

Taking a step back, recording all changes in the BOM is usually not
significant. Let's consider the following BOM modifications:
- Add a product
- Modify the product quantity
- Remove the product

If the user saves after each step, 3 BOM Changes are recorded. If the
user saves only at the end, no change is recorded. Recording all
modifications will likely bloat the differences between the two BOMs.

Therefore, we clear all computed fields which use the same logic in
order to record only the global differences.

opw-2178950

closes odoo/enterprise#7965

X-original-commit: bbf0afe23a87892502992825e50e29fa67e03002
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>