[FIX] stock_barcode: scan a reserved package
On a delivery with move entire package enabled, a package is
reserved. If you scan this package, it is highlighted but its processed
quantity is displayed as 0. If you validate the picking, everything is
moved though.

The issue was that the reserved move lines were not incremented due to a
bad condition but new ones were created instead.

opw-2171806

closes odoo/enterprise#7948

X-original-commit: 9e2bdc5c02a919811ae8cca6e88146dda75e7a4c
Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>