[FIX] account_reports: Fix unaffected earnings in Trial Balance Report
Suppose two invoices of 1000.0 respectively in 2017 & 2018.
- Open the Trial Balance in 2018:

An initial balance of 1000.0 is reported in the current year earnings account
corresponding to the invoice in 2017.

- Set a comparison with the previous period.

The current year earnings account is reported with 0.0 amount everywhere.

Explanation: As the unaffected earnings was computed for each period but reported only for
the oldest one, the current year earnings account was involved by 2018 but only the
initial balance of 2017 (0.0) was reported.

This fix will only computed the unaffected earnings for the oldest period and then, will
display the current year earnings account only according this one.

closes odoo/enterprise#7981

X-original-commit: 86e1994ec8ec6bd1c436d9dd0cc65b01b4c22b7a
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>