[FIX] account_reports: unlinking an "account.financial.html.report"
The previous fix (eef83eecca) was not 100% working and some menus and
actions were still dangling after unlinking a deprecated report.
One of the reason is that some menus are not directly under the parent
menu but one level below.

The easiest way to unlink those menus and actions is to use the
'generated_menu_id' field. This field points to the menu that is
programmatically generated by the '_create_action_and_menu' method.
This way, we are 100% sure that the menu and action related to the
deprecated report are also deleted.

opw:2169636

closes odoo/enterprise#7960

X-original-commit: 1627cdc73865ed8fe6ce2daab259ea0cab5a8b8d
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>