[IMP] account_asset: allow to use asset with already posted entries
Task 2182787

When you start using Odoo, older data should be managed.
Assets already booked for closed fiscal years should be well reported
through the depreciation report.
This task aims to easily Create them accordingly.

While we are at it, we also use the changes done in the ORM recently
with the editable compute. This makes the behavior more stable and the
code more understandable.

closes odoo/enterprise#7972

Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>