[IMP] social_facebook: view event creation post on the feed
Purpose
=======

On Facebook, when we create an event, a post will be automatically created.
Before, those posts were empty but we want to display them in the feed view.

Specifications
==============

Retrieve the link to the event and the image of the event to display it in
the feed view (the link of the event is displayed as a simple link,
and the cover of the event is displayed as the post images).

Task ID 1999230
Closes PR #8058