[ADD] event_social_facebook: add Facebook events support
Purpose
=======

Post on Facebook event and link "Event" from the "Event" application to
Facebook

Specifications
==============

We want to be able to post on the events on Facebook.

We also want to link the Facebook events, to the events in Odoo.
Display the Facebook statistics in the event kanban views, post on
Facebook from the event form view, fill automatically the event
information with the information on Facebook.

API Limitations
===============

On the date of the 28 January 2020, we have some API limitations which are
the following

  * cannot create event with the API. Events must be created in facebook
    then synchronized within Odoo;
  * cannot read the feed of an event. Statistics are extracted and posts
    are retrieved but reading event itself is not possible;

Task ID 1999230
Closes PR #8058