[REF] quality, mrp_workorder, *: point no more product template
The configuration for variant and template are difficult to understand:
The point will apply for the specific variant if they are define.
However it will apply if only the template is defined but there are no
variant explicitely specified. It's easier to have a simple
configuration as the product variants specified are executed.