[IMP] web_studio: adapt kanban editor to auto kanban
The ModelConfigurator flow adds automatic kanban view creation, which
generates a 'complex' kanban view for new models automatically.

This means that the hooks to add automatic behaviours in the kanban
editor had to move a litte or to hide in different conditions to avoid
inserting widgets that look too much alike (e.g. boolean_favorite and
priority widgets).