[IMP] mrp: Add new components in workorders
Purpose: adding new component in a specific workorder via the form view.
This behavior is only available if the Bill of Material is in flexible
consumption. Adding new component was already made possible in
https://github.com/odoo/odoo/commit/e1f518bb83ccaa2abf08bc870bc1c7345e85971c
but the ability to add new workorder lines conditionnaly in the workorder
form view was missing.

closes odoo/enterprise#7253

Task: 2121714
Related: odoo/odoo#41631
Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>