[FIX] delivery_ups: cancel delivery of several packages
When several packages are delivered in a delivery order, if we cancel
the delivery we have to cancel each package one by one because UPS only
has API to cancel only one shipment.

This was not done so we would get an error.

With this changeset, we will cancel 5 packages if there is 5 package on
the current delivery order.

opw-2180729
closes #8016

closes odoo/enterprise#8095

X-original-commit: 4fa58ef850757bbca2120cdbc2d414034e015074
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>