[FIX] account_reports: 'allowed_company_ids' should always be there in context
In multicompany scenarios, when we access company property of environment,
that is, self.env.company, from within the methods such as get_xml, get_xlsx
etc. when they are called from account_reports controller, we always get current
user's company in it. It should instead give us company that we have selected
in company switcher.

closes odoo/enterprise#8065

Task: 1952265
X-original-commit: c6326960b10c163dd55598202a6c2997389ebd0e
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>