[IMP] project_forecast: Undefined User issue in the Gantt view
Before commit:
In project current group by on Gantt view is on 'user_id' not by
the employee_id so, when you change the employee and that employee
is not linked with the user then it will display the 'Undefined User'.

After commit:
We change the group by on 'employee_id' so, in gantt view display the
'open shift' as well as display the corresponding 'employee' under the
open shift.

closes odoo/enterprise#8057

Task-id: 2172834
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>