[IMP] l10n_mx_edi: Do not allow to delete a payment signed in the SAT
In the next case:
1. Generate an Mexican invoice and sign it.
2. Make the invoice payment with a bank statement and generate the
payment complement.
3. Un-reconcile the bank statement.

In this moment Odoo, remove the move in the payment generated from the
bank statement and unlink the record the payment.

In
https://github.com/odoo/enterprise/commit/88c2666fd7a375e6cf7e97ec5747c0549439fb1f
was improved to avoid unlink a payment if the PAC status is ``signed``, but
in some cases the payment cannot be cancelled, and the PAC status is
``to_cancel``. For this reason, now in the unlink process is updated the
SAT status and only allow unlink if the CFDI is cancel in the SAT.

Cases:
1. Payment complement was not generated, maybe the bank statement is an
   advance and the payment generated do not have invoices related.
   In this case, the payment do not have CFDI and allow remove without
   problem.
2. Payment complement generated but not was signed.
   When the PAC status is in 'to_sign' or 'retry' and is called the
   method ``_l10n_mx_edi_cancel``, automatically the
   ``l10n_mx_edi_pac_status`` is False and allow unlink the CFDI.
3. Payment signed, try unlink but the payment cannot be cancelled.
   In this case, the SAT status must be ``to_cancel``, and the SAT
   status ``Valid``, then the payment cannot be unlink.
4. Payment signed, cancel correctly.
   In this case the payment could be removed.

Note: In test instances is not possible get the SAT status = cancelled,
for this reason, in test is omitted this validation.

closes odoo/enterprise#8292

X-original-commit: 150c5eebde6269e921e057eef3ba3fe59b5362e4
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>