[IMP] social: display a quick create form view when creating campaigns on the fly
When creating a new post, interface allows to directly create a new utm
campaign linked to the post. Display form view is the default one with
notably action buttons and state bar, which clutter the interface.

In this commit we introduce a small form view allowing a nice and user friendly
quick create.

Task ID 1999230
Closes PR #8058

Related: odoo/odoo#44359
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Co-authored-by: Aurélien Warnon <awa@odoo.com>
Co-authored-by: Thibault Delavallée <tde@odoo.com>