[IMP] L10n_be_reports: Improve generating of 281.50
This commit adding some new checks and features for the generating of 281.50 slip.
Now, there is a new partner tag named "281.50". This tag is mandatory if the
user wants to generate a 281.50 slip for a specific partner.
For the selection of the year in the initializing wizard, the period constrain
is removed but if you select a year superior or equal to the current year,
the slip will be generated in "test" mode.
Access Right: Moving from Advisor to Accountant.

Task ID: 2186378

closes odoo/enterprise#8061

Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>