[FIX] L10n_be_reports: Fix SQL queries for paid amount.
Problem fixed:

For partner X
Invoice of 121€ -> 100€ + VAT 21€ in 2018.
Pay 100€ in 2018 and 21€ in 2019.
Invoice of 121€ --> 100€ + VAT 21€ in 2019.
Pay 100€ in 2019 and 21 in 2020.

For partner Y
Invoice of 121€ -> 100€ + VAT 21€ in 2018.
Pay 100€ in 2018 and 21€ in 2019.

Generate 281.50 slip for partner X. Error is raised.
This error is raised because the sql queries in method
to compute "paid amount" didn't take account of partner ID.
Thus, when you wanted 281.50 slip for a specific partner,
we tried to get the currency of partner Y. But we didn't provide this
partner in recordset.

closes odoo/enterprise#8083

X-original-commit: 745f6d020ec30d1df8d3e361f30cf7995e8a412c
Signed-off-by: Florian Gilbert <FlorianGilbert@users.noreply.github.com>