[FIX] stock_barcode: wrong readonly=False on product_barcode
Usecase to reproduce:
- Set show details operation on a picking type.
- Connect as demo user
- Create an immediate transfer and add a line with quantity_done
- Save

It will raise an access error since you are not allowed to write on
product.product. However it's not the case we only created a move line
and we won't write any information on product. It's due to a related
field product_barcode on stock.move.line that is set as readonly=False.

So when the move line is created it will try to write the barcode on the
product.

closes odoo/enterprise#8094

X-original-commit: 658234d4e791a9977bf8700352ffe8de3bf25bbf
Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>