[IMP] planning: Don't erase group_expand if #employees <= 20
Purpose
=======

Most of our customers are SMEs with a handful employees.
For these people, it is practical to have a list of all
employees always available.

However, this could cause significant performance issues for
big databases with a lot of employees. In that case, the
current group_expand we have would be perfect, as it only
shows the essential information.

Specification
=============

If there are less than 20 employees (or any other number
that suits you best), display all employees in the Gantt view

If there are more than 20 employees, only display the employees
to whom a shift was assigned in the past 30 days (current
group_expand we have in master)

closes odoo/enterprise#8967

Taskid: 2184024
X-original-commit: c7f7bb3d017af4c581f55bca568e665657de15dc
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>