[FIX] website_enterprise: incompatibility with website_livechat
website_enterprise hooks itself to `module_website_livechat` to know
what the last-ish menu item is in order to add
`module_website_calendar`. However website_livechat kinda removes that
node (in that it removes the `name` and `data-module-id` attributes of
the element). As a result if the website_enterprise extension is
applied after the website_livechat one, the former blows up.