[FIX] industry_fsm_sale_report: inheritance of wrong view
industry_fsm_sale_report is a bridge module for industry_fsm_sale and
industry_fsm_report.

Both define a `view_task_form2_inherit` which the bridge wants to
alter somewhat. It extends the version from sale, however what it does
really hooks into is the report one, as it updates
`button[@action_send_report]` (which doesn't sound much like sales
stuff).

I'm not sure it's possible for the issue to occur in standard odoo (as
modules are loaded in lexicographic order when all their dependencies
are fulfilled, and industry_fsm_report is both ordered before
industry_fsm_sale and has a simpler and shallower set of dependencies)
but it seems prudent not to rely on implementation details.

closes odoo/enterprise#8098

Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>