[FIX] web_enterprise: disable tour in mobile
Tours were activated in tour_services.js;
See https://github.com/odoo/odoo/pull/46880

But we have to disable it on mobile. It's why we automatically
consume all tours to avoid to display it for demonstration
purposes.

This code is already used to disable tours if Odoo is installed
with demo data.

Task ID: 2149402