[FIX] account: correct blur event when editing amount in reconciliation widget
Commit: https://github.com/odoo/enterprise/commit/a1f6cf1367d254a6515546599ac624a823916890
changed the event on input from keyup,blur to change which was wrong.
Change event only trigger when an input value has changed, or in our case when we edit the
amount of a line, we want the input to disappear when we click elsewhere or when we press enter
and no change were made to the value. This commit fixes this by reintroducing the keyup and blur
event.

closes odoo/enterprise#8120

Signed-off-by: Cedric Snauwaert (csn) <csn@openerp.com>