[FIX] planning: working days shouldn't be zero
Before this commit, the working days count was only filled if the
planning number of hours was bigger than the hours working day. If an
employee is planned to work two hours in the task, it should be one
working day and not zero as it's the case for the moment.

Now, the working days are rounded up.

opw-2178341

closes odoo/enterprise#8155

X-original-commit: c30a9eee2a864185538d6b98ac3b9ab1851ec8f5
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>