[REM] *: remove various unused import shims
Before this commit, a lot of leftover import shims existed in the
codebase either for py2-py3 compatibility or for soft dependencies that
are now hard dependencies and in one particular case a third-party
module dependency was merged into the CPython codebase (making the
dependency useless).

With this commit, these shims are dropped, making the code cleaner,
easier to read and with one less dependency.

closes odoo/enterprise#8141

Related: odoo/odoo#44601
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>