[FIX] social_demo, event_social_facebook: fix some Facebook related issues
Bug 1

When `social_demo` is installed, we can not post on Facebook as an override
of _post_facebook has not the same parameters as the base method.

Bug 2

When creating a new account in event social facebook tests, some default
streams are automatically created which make tests crash. We should mock the
method `_create_default_stream_facebook` to avoid that.

Task 2188857

closes odoo/enterprise#8533

X-original-commit: 0f4a2fc44c6fb4ebc501c7cba4c5c73a4e3d6f23
Related: odoo/odoo#45655
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>