[FIX] account_asset: deny move reset to draft on opened deferred rev/exp
Create a deferred revenue/expense
Create an invocie which map to the deferred revenue/expense account
Post the invoice so it will generate the deferred entry (in draft)
Confirm the deferred Entry
Reset to draft the invoice

User will be unable to post the invoice again because of the deferred
revenue/expense open but also unable to access it because the invoice is
not in posted state (deadlock).

Adding a condition to block 'reset to draft' button if there is some
deferred revenue/expense not in draft.

opw-2180330

closes odoo/enterprise#8494

X-original-commit: a3cf1d8d5b4316fc61e5656995023453727d6c5e
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>