[FIX] Timesheet_grid: Avoid make group_expand with date
Before this commit, when we try to make a group by date in timesheet
grid, we have a traceback. This group by date is useless in grid view
but very usefull for list view.

Now, when we make a group by date in the timesheet grid, we don't
apply this group_by.

closes odoo/enterprise#8194

Taskid: 2169272
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>