[FIX] payment_sepa_direct_debit: correct payment method
Without the `payment_method_id` field of the linked account.payment
record being set to the SDD payment method, it was impractical (or
impossible, really) for users to export payments generated that way in
an xml file to send to their bank.

By using this payment method, we ensure that the available payments will
be listed in the accounting dashboard and will be exported easily.

closes odoo/enterprise#8251

X-original-commit: a1b5bba3310e7100294f92d685352cea3e04435c
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>