[FIX] delivery_ups: cancel delivery of several packages (bis)
In 3b1d94cbe5a all reference of a package would be tried to be
cancelled, so the cancelling would would be with multiple reference we
would have error on first cancel:

REF2: No shipment found within the allowed void period
REF3: No shipment found within the allowed void period

then on second cancel:

REF1: The Shipment has already been voided
REF2: No shipment found within the allowed void period
REF3: No shipment found within the allowed void period

and because we send an error, it was not clear that the first request
worked.

With this changeset, we go back to previous behavior and only cancel the
first reference which is what UPS API is expecting.

opw-2180729
closes #8205

closes odoo/enterprise#8255

X-original-commit: c160599bf41303ed52aa2b47a1f0f3ed07acf091
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>