[IMP] industry_fsm_sale : make the value editable on product kanban view
Material view
=============
When industry_fsm_sale is installed, the user as the
opportunity to add materials to his tasks. This is done
with a stat button that shows a kanban view of products
where he can add the materials.

Previously, if he wanted to add 50 units of something
he add to click on it 50 times.

This wasn't practical so a widget has been added to the view
to allow the user to manually edit the value on the kanban
view.

Task-2188831

Co-authored-by: jdoutreloux <jud@odoo.com>