[IMP] industry_fsm_sale: display quantity buttons on product
Purpose of the task is that many users are confused when they reach
the products kanban view from the 'Add products' stat button and
don't understand what they are expected to do.

so display the quantity buttons alaways in the product kanaban view
if quantity is 0 then buttons should be muted.

Task-2170719

closes odoo/enterprise#8256

Closes: #8256
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>