[IMP] web_{gantt,grid}: evaluate ACL as current user
Apply same patch as in super method _postprocess_access_rights
The view content is not accessible by non-priviledged users so the
method fields_view_get is executed with a super-user environment.
The _postprocess_access_rights evaluates if the current user can or
can not create/write/delete a record to display the appropriate
buttons.
This method must be executed without the su to avoid misleading the
user and displaying a button they won't be able to use.