[FIX] account_accountant: fiscal year end dates in settings
- From Accounting Settings > Fiscal Periods > Fiscal Year;
- Change fiscal year last day from December 31 to November 30;
- Click Apply button

An error is raised: 'Invalid fiscal year last day'.

This is because the 2 fields (Month and day) are linked by the
constraint `_check_fiscalyear_last_day`. However, non-stored writeable
related fields are evaluated one at a time, meaning that the constraint
is evaluated one value at a time.

We do a fancy workaround to write the 2 values at once on the company,
hence evaluating the constraint for the 2 fields once.

Same fix as odoo/odoo@988018a5baa1a10ee44ea

opw-2187056

closes odoo/enterprise#8434

X-original-commit: 931b88fa38764263bf05b1b495deec70b3459b44
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>