[IMP] social(_*): split `social.post` into `social.post.template`
Purpose
=======
We want to be able to write social posts, without the intention to send
them directly. For that purpose, we moved some fields of `social.post`
to a new model, `social.post.template`. The model `social.post.template`
has the common fields of both, mainly fields related to the post content
(so, be inheriting, we avoid code duplication).

Move also the `account_ids` fields, so we can write template depending
on the media on which we will send.

So, in Social Event, we are able to generate multiple social posts from
the same social post template.

Specifications
==============
All fields not related to the publishing has been moved to
`social.post.template`.

As the preview of the post is also moved to `social.post.template` we
can no more use the scheduled date in the preview and we simply use the
current date.

Links
=====
Task-2127615
See odoo/odoo/pull/46304
See odoo/enterprise/pull/7701
See odoo/upgrade/pull/848