[FIX] l10n_fr_hr_payroll: make module installation possible
When installing the module, there was a KeyError for model
hr.leave.type as there was an attempt for it's record insertion.
The module has no direct/indirect dependency for hr_holidays
module.

Currently the record is removed. There was an alternate for
adding hr_work_entry_holidays module dependency, but the leave
type with string 'Heures Supplementaires' seemed inapt. So the
record was removed.

Certain created records getting updated later had missing model
attribute, so those were filled.

closes odoo/enterprise#8442

Task: 2169189
X-original-commit: be28fc3f1cde6cb6cf5241d7fb11e1cad69dff1f
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>