[FIX] inter_company_rules: create SO fro PO with section
- Create Company A & B
- Activate 'Synchronize sales/purchase orders' for A & B
- In A, create an RFQ for B with a section or a note
- Validate

A `ValidationError` is raised: `accountable_required_fields`.

This is because the `display_type` is not propagated from the PO to the
SO.

We can propagate the value as well as stop filtering the lines.
Filtering was introduced was sections and notes were not supported in
PO. This is not the case anymore.

opw-2195653

closes odoo/enterprise#8430

X-original-commit: fb72b16ed2cfe64252b13d7f253f686b0cd858a9
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>