[FIX] quality, quality_control: slip redirection
- Create a BOM with product A and component C
- Go to Quality / Quality Control / Control Points
- Create 3 QC for A
- Create a MO for A, e.g. of type 'Measure'
- Click on the stat button 'Quality Checks'
- Pass or Fail the first test

A pop-up appears asking for QC of the second test.

The redirection is done because of the following override:

https://github.com/odoo/enterprise/blob/88265a07b6b754cc8fb3d5885c371f5b52933a5d/quality_mrp_workorder/models/quality.py#L17

The redirection makes sense when the QC are launched through the action
button 'Quality Checks' on the MO. Indeed, in this case the pop-up
directly opens. However, this is not useful from the regular QC form
view and moreover, it is confusing from the end user point of view.

We skip any redirection in the regular QC form view.

opw-2186963

X-original-commit: 7eea4608bfb9a03d8553145bdee9435f3ca4b314