[FIX] quality_control: add 'Measure' to QC form view
When using the regular QC form view, a user can make pass or fail any
measure test.

We add a 'Measure' button to validate the measure. Note that the usual
'Pass' and 'Fail' buttons are still available, and the user can still
pass or fail any measure if he wants.

opw-2186963

closes odoo/enterprise#8429

X-original-commit: e84af9065b439b8f2ad15b949d62cb5c8d091275
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>