[ADD] account_reports_cash_basis: cash basis for financial reports
Task 2198748
The functionality was removed by https://github.com/odoo/enterprise/pull/4379
because the functionality was buggy, laggy, and since we didn't get a
lot of tickets about it, we thought it wasn't used much. As it turns
out, it was.

This commit reintroduces the functionality while improving the
performances and hopefully removing bugs.

closes odoo/enterprise#8970

X-original-commit: 8e997d92a825a0b7ed1603a6e8da6201940eea8f
Signed-off-by: wan <william-andre@users.noreply.github.com>