[IMP] documents: enabling studio should not throw traceback in documents
before this commit, enabling studio in documents kanban view throws traceback,
this is because in studio mode searchPanel will not be rendered and many2many tags
which are displayed on kanban record are depends on searchPanel due to which for..of
loop will throw error as this.tags will be undefined and not iterator.

after this commit: enabling stuido will render tags like normal kanban m2m tags,
here we will have bit different UI as we don't have searchPanel available in studio
so we will not have proper colors and tags UI like we have in normal kanban view in
documents but it is better to not have traceback, we can compromise with UI for
studio mode.

task-2189992

closes odoo/enterprise#8441

Signed-off-by: Damien Bouvy (dbo) <dbo@odoo.com>