[IMP] web_studio: display chatter below the sheet in studio mode
before this commit: when we enable studio then chatter comes in right side for
bigger screens, chatter takes 25% width of the screen so for editing form view
we have 75% of the screen, so user has less space to edit form view.

after this commit: if studio mode is enabled then chatter will always be at
bottom of the form regardless of the screen size, this allows user to use more
space to edit form.

task 2183697

closes odoo/enterprise#8550

Signed-off-by: Simon Genin (ges@odoo) <SimonGenin@users.noreply.github.com>
Co-authored-by: Mohammed Shekha <msh@odoo.com>
Co-authored-by: Yograj Tandel <yta@odoo.com>