[IMP] hr_referral: Add not hired stage
Add a boolean on recruitment stage to said that is a 'not hired stage'.
When it's true:
1. this recruitment stage will not be display on
referral dashboard.
2. this stage give no points (neither if this stage is at the end of
process, and we go for example from stage seq 4 to this stage with
sequence 10, no points will bre gived). And after, if we go to normal
stage with sequence 6, we will received points for stage 5 and 6.

closes odoo/enterprise#8491

Taskid: 2165770
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>