[REF] web_mobile, mail_mobile: auto install and move settings
Before this commit, mail_mobile (ocn_client) was installed if you checked
a setting added by web_mobile. Once installed, notifications worked (if
project_id was set).

But now, we have to install it when mail_enterprise is installed because
this module will contains multiple features that may be activated like
"URL redirect to mobile app".

After this commit, mail_mobile will be automatically installed when web_mobile
and mail_enterprise are installed but for now we don't want to activate
the notification. It's why we moved this settings from web_mobile to mail_mobile.
This setting will not install any module anymore but only activate this feature.

Task ID: 2178953

X-original-commit: c13f1d82ee52079648a5ed4c4ac832e2bfddac14