[MOV] web_mobile, mail_mobile: move redirect feature to mail_mobile
We have to do this in 13.1 because of the migration tests failure.
There was a missing dependency in web_mobile: "mail".

As we don't want to add this one in web_mobile, we have to move it to
another module: mail_mobile that already depends of mail_enterprise.

Task ID: 2178953

X-original-commit: add08473cf05faf9432842380377ae87e8581c24