[FIX] web_studio: fix wrong x2m z-index overlay
before this commit: when x2m field is clicked it shows overlay on field
and that overlay has z-index 1100 due to which if dialog is opened on top
of that then dialog elements go behind that overlay, for example x2m overlay
is available and if conditional dialog is opened from sidebar in studio mode
then conditional domain dropdown goes behind x2m overlay which is wrong.

after this commit: x2m blocked overlay will have z-index 1000 so that dialog
or any other elements will not go behind x2m overlay.

task-2187354

closes odoo/enterprise#8521

Signed-off-by: Damien Bouvy (dbo) <dbo@odoo.com>