[FIX] account_invoice_extract: no duplication of box layers
Before this commit, at each thread re-rendering a box layer is created
if the uploaded file is an image. When creating the box layer, the old
one is replaced by the new one but the old one is not removed from the
DOM. It does not cause any functional problem as you only interact with
the last created layer, so the other ones are not accessible and useless
and as they are transparent they don't cause any problem. By the way,
it's not a good thing to have useless DOM elements.

This commit fix this problem by removing all the previously created
layers from the DOM before replacing them with the newly created one.

closes odoo/enterprise#8542

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>