[IMP] helpdesk_sale_timesheet : add a Create Invoice button on ticket
The user can now create a invoice from a ticket via a button on the
ticket itself.

When the SO is invoiced or cancelled, the Create Invoice button
doesn't appear. If the SO invoice status is upselling or to invoice
the button is primary. If it's nothing to invoice, the button
is secondary and the default payment method is downpayment.

Task-2191253

closes odoo/enterprise#8530

Related: odoo/odoo#45650
Related: odoo/upgrade#808
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>