[IMP] account_reports: tax report grouped by GL account
Task 2194008
VAT Summary by General account  allows you to print for each G/L account
the list of VAT codes with the total of the bases and the associated VAT
amounts.

This list serves among other things to help you draw up Annex A of the
Luxembourg annual VAT return but can, for example, be useful when it
comes to distinguishing by country the VAT to be reclaimed abroad or for
VAT Auditing.
It’s mandatory for Luxembourg but it's still useful for generic auditing
so we can add it globally

closes odoo/enterprise#15111

Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>
Signed-off-by: wan <william-andre@users.noreply.github.com>