[IMP] l10n_be_hr_(contract_salary, payroll_account): Change the employee private address as 'private'
Purpose of this task is,lots of business flows can be crashed if the field
Private Address(address_home_id) is a private address instead of a regular
contact.

so in this commit,we change the demo data for employee and set the
private type address on employee for private address and also fix
the flows on which errors could occur.
Changed lap address as private address and in the test module changed
all the address as private address instead of creating the new for each
employee.

closes odoo/enterprise#8924

Taskid: 2170016
Closes: #8924
Related: odoo/odoo#46628
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>