[FIX] account_winbooks_import: reconcile through multiple years
* If we try to import the accounting entries through multiple files (i.e.
one per fiscal year), we might encounter this case:
3 lines are reconciled; the first 2 are in one file and the last one in
another file. Because we try the reconciliation at the end of every
import, we will have one of the first 2 lines that is marked as reconciled,
the other still with an open amount.
When we import the second file, we are trying to reconcile all the 3
lines because they had the same matching number in winbooks, but it will
raise an error because one line is already reconciled.

We should only take the line that was with an open amount to reconcile
it with the third line.

* Another issue is that taxes with the same name might change the amount
through the years.

closes odoo/enterprise#8702

X-original-commit: ff5718aef62790044b047c673b93ca022e08fd5b
Signed-off-by: Cedric Snauwaert (csn) <csn@openerp.com>