[IMP] web,timesheet(_grid): Several improvement on grid view.
1) Currently, when the whole grid view is readonly(create/edit False) then
it will allow to create a analytic line from magnifying icon and allow
to import the analytic line.

So After this commit if the whole grid view is readonly then do not
allow to edit/create analytic line from the list view which is opened
by magnifying icon click.

2) Currently, when the grid view is in read-only and the there is no
analytic line linked with that cell, the magnifying glass icon opens
a list view which is completely useless as it is empty.

So after thid commit, magnifying glass icon will not display for the
readonly cell which has no analytic line linked with that cell.

3) Currently, when creating or editing value in grid view, then
read_grid method is called and if line is not present then create a
new one by copying previous line which has similar domain, because of
that validated field was also copied from there so default new line is
going to be validate while it's not.

So After this commit we make validated field as Copy=False so when
the analytic line is copied it will not validated by default.

TaskID: 2184268
Closes: #8927