[IMP] stock_barcode_picking_batch: remove picking from barcode view
As a batch picking can't be validate if one of its picking still in
waiting state, user needs a way to unlink a picking from the batch in
case he/she wants to validate the batch anyway.
It's not a big deal in form view as the user has already the picking
list and can use it to remove faulty pickings, but it's not as easy in
the batch picking barcode view.
So, this commit adds pickings list in the informative view (which user
can acces to by click in the 'i' icon in the barcode view), set faulty
pickings in red and adds a button in this list to remove these pickings.

task-2123726