[IMP] sign: Allow to save document in progress
Now, when we sign a document, we can save the document in progress.
The next time that we open this signature request, all saved
informations will be fill. Other signatories won't see signature in
progress (only visible when signatory has sned and validated the
document.

And allow to rotate pdf

taskID: 2196770