[FIX] stock_enterprise: fix traceback on grouping error
Inventory>Reporting>Forecasted Inventory
Group by something other than product (i.e. Warehouse)
Click on the magnifying lens in a cell

Traceback will occur because frontend and backend expect to
deal always with product_id, instead the supplied data when grouping is
relevant to the grouping subject, so will crash the interface.
Adding an error display message to inform the user of the problem.

opw-2199757

closes odoo/enterprise#8730

X-original-commit: 828fa3bf6cefe213ad3c71beae483115f0ab1d60
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>