[FIX] hr_expense_extract: avoid to change already set product
Before this commit:
In expense, if we change the name of an expense, the product and
the price will be overridden by a new default one (comes from the
predict of product based on name). But it makes non sense to change
already set product when we change the name (for a typo for example)

After this commit:
We try to predict product when we change the name if they are no
already set product.

closes odoo/enterprise#9138

X-original-commit: b00e9b43910d8b30e463b2430f5187d26fc241c2
Signed-off-by: jbm-odoo <jbm-odoo@users.noreply.github.com>